created generic test generation
