+++ app/lucy/src/app/components/base-form/base-form.component.ts
@@ -7,22 +7,19 @@ import {
   Renderer2,
 } from '@angular/core';
 import { FormMode } from 'src/app/models';
-import { ErrorService, ErrorType } from 'src/app/services/error.service';
 import { UserService } from 'src/app/services/user.service';
 import { RolesService } from 'src/app/services/roles.service';
 import { AlertService } from 'src/app/services/alert.service';
 import { RouterService } from 'src/app/services/router.service';
 import { UserAccessType } from 'src/app/models/Role';
-import { AppRoutes, AppConstants } from 'src/app/constants';
-import { DropdownObject, DropdownService } from 'src/app/services/dropdown.service';
+import { AppRoutes } from 'src/app/constants';
 import { FormService, FormSubmissionResult, UIConfigObject} from 'src/app/services/form/form.service';
 import * as moment from 'moment';
-import * as faker from 'faker';
-import { ApiService, APIRequestMethod } from 'src/app/services/api.service';
 import { LoadingService } from 'src/app/services/loading.service';
 import { DiffResult } from 'src/app/services/diff.service';
 import { ElementRef } from '@angular/core';
 import { ToastService, ToastIconType } from 'src/app/services/toast/toast.service';
+import { DummyService } from 'src/app/services/dummy.service';
 
 export enum FormType {
   Observation,
@@ -155,11 +152,11 @@ export class BaseFormComponent implements OnInit, AfterViewChecked {
   }
 
   // Config
-  private _config: any = {};
-  private get config(): any {
+  private _config: UIConfigObject;
+  private get config(): UIConfigObject {
     return this._config;
   }
-  private set config(object: any) {
+  private set config(object: UIConfigObject) {
     this._config = { ...object};
   }
 
@@ -227,15 +224,12 @@ export class BaseFormComponent implements OnInit, AfterViewChecked {
   }
 
   constructor(
-    // private mechanicalTreatmentService: MechanicalTreatmentService,
-    private errorService: ErrorService,
+    private dummy: DummyService,
     private userService: UserService,
     private roles: RolesService,
     private alert: AlertService,
     private router: RouterService,
-    private dropdownService: DropdownService,
     private formService: FormService,
-    private api: ApiService,
     private loadingService: LoadingService,
     private elementRef: ElementRef,
     private renderer: Renderer2,
@@ -299,9 +293,12 @@ export class BaseFormComponent implements OnInit, AfterViewChecked {
     // if input was invalid, field component emits ``
     // handle INVALID input cases
     if (field.isLocationField && (event.latitude.value === `` || event.longitude.value === ``)) {
+      // console.log('setting lat long in body to undefined')
       this.responseBody[field.latitude.key] = undefined;
       this.responseBody[field.longitude.key] = undefined;
     } else if (event === `` && this.responseBody[field.key] !== undefined) {
+      // console.log(`setting ${field.key} to undefined`)
+      // console.dir(event);
       this.responseBody[field.key] = undefined;
     }
     // handle valid input cases
@@ -344,6 +341,7 @@ export class BaseFormComponent implements OnInit, AfterViewChecked {
     // const endpoint = `${AppConstants.API_baseURL}${this.config.api}`;
     if (!this.canSubmit) {
       this.triedToSubmit = true;
+      console.dir(this.responseBody);
       this.toast.show('Some required fields are missing', ToastIconType.fail);
       return;
     } else {
@@ -435,149 +433,11 @@ export class BaseFormComponent implements OnInit, AfterViewChecked {
     }).replace(/\s+/g, '');
   }
 
-  async generateTest(configuration: UIConfigObject): Promise<UIConfigObject> {
-    const config = { ...configuration};
-    for (const section of config.sections) {
-      for (const subsection of section.subSections) {
-        for (let i = 0; i < subsection.fields.length; i++) {
-          const field = subsection.fields[i];
-          if (field.isComputedField) {
-            continue;
-          }
-          if (field.isLocationField) {
-            const fakeLat = await this.fakeFieldValue(field.latitude);
-            field.latitude.value = fakeLat.fieldValue;
-            this.responseBody[field.latitude.key] = fakeLat.bodyValue;
-            const fakeLong = await this.fakeFieldValue(field.longitude);
-            field.longitude.value = fakeLong.fieldValue;
-            this.responseBody[field.longitude.key] = fakeLong.bodyValue;
-          } else {
-            const fake = await this.fakeFieldValue(field);
-            field.value = fake.fieldValue;
-            this.responseBody[field.key] = fake.bodyValue;
-          }
-        }
-      }
-    }
-    console.dir(config);
-    console.dir(this.responseBody)
-    return config;
-  }
-
-  private async fakeFieldValue(field: any): Promise<{fieldValue: any, bodyValue: any}> {
-    if (field.isDateField) {
-      const date = faker.date.past();
-      const fake = moment(date).format('YYYY-MM-DD');
-      return {
-        fieldValue: fake,
-        bodyValue: fake,
-      }
-    }
-
-    if (field.isCheckbox) {
-      const fake = faker.random.boolean();
-      return {
-        fieldValue: fake,
-        bodyValue: fake,
-      }
-    }
-
-    if (field.isTextAreaField) {
-      const fake = faker.lorem.sentences();
-      return {
-        fieldValue: fake,
-        bodyValue: fake,
-      }
-    }
-
-    if (field.isInputField && field.type.toLowerCase() === 'string') {
-      const fake = faker.lorem.word();
-      return {
-        fieldValue: fake,
-        bodyValue: fake,
-      }
-    }
-
-    if (field.isInputField && field.type.toLowerCase() === 'number') {
-      const fake = this.randomNumber(4, 20)
-      return {
-        fieldValue: fake,
-        bodyValue: fake,
-      }
-    }
-    
-    // Dropdown
-    if (field.isDropdown) {
-      const randomIndex = this.randomNumber(0, field.dropdown.length - 1);
-      const value = field.dropdown[randomIndex];
-      const vieldValue = value;
-      let selectedID = 0;
-      for (const key in value) {
-        if (key.toLowerCase().indexOf(`id`) !== -1) {
-          selectedID = value[key];
-          break;
-        }
-      }
-      return {
-        fieldValue: vieldValue,
-        bodyValue: selectedID,
-      }
-    }
-
-    if (field.key.toLowerCase() === `lat` || field.key.toLowerCase() === `latitude`) {
-      const value = this.randomLat();
-      return {
-        fieldValue: value,
-        bodyValue: value,
-      }
-    }
-
-    if (field.key.toLowerCase() === `lon` || field.key.toLowerCase() === `long` ||field.key.toLowerCase() === `longitude` ) {
-      const value = this.randomLong();
-      return {
-        fieldValue: value,
-        bodyValue: value,
-      }
-    }
-
-    console.log(`Unknown field ${field}`);
-    console.dir(field);
-    return {
-      fieldValue: undefined,
-      bodyValue: undefined,
-    }
-    
-  }
-
-
-  /**
-  * Generate a random longitude within bc
-  */
-  public randomLong(): number {
-    const a = this.randomNumber(0, 7);
-    const b = this.randomNumber(845602, 977180);
-    const z = `-12${a}.${b}`;
-    return +z;
-  }
-
-  /**
-   * Generate a random latitude within bc
-   */
-  public randomLat() {
-    const a = this.randomNumber(0, 8);
-    const b = this.randomNumber(713134, 202679);
-    const z = `5${a}.${b}`;
-    return +z;
-  }
-
-  private randomNumber(min: number, max: number): number {
-    return Math.floor(Math.random() * (max - min + 1) + min);
-  }
-
   async generateForTesting() {
     this.loadingService.add();
-    this.config = await this.generateTest(this.config);
-    console.dir(this.config);
+    const fake = await this.dummy.generateTest(this.config);
+    this.config = fake.config
+    this.responseBody = fake.json;
     this.loadingService.remove();
   }
