+++ app/lucy/src/app/components/base-form/location-input/location-input/location-input.component.ts
@@ -49,17 +49,29 @@ export class LocationInputComponent implements OnInit {
 
     if (this.object && this.object.latitude && this.object.latitude.value) {
       // console.log(`setting ${this.object.latitude.value}`);
-      this.lat = String(this.object.latitude.value);
+      this.lat = `${this.object.latitude.value}`;
     }
     if (this.object && this.object.longitude && this.object.longitude.value) {
       // console.log(`setting ${this.object.longitude.value}`);
-      this.long = String(this.object.longitude.value);
+      this.long = `${this.object.longitude.value}`;
     }
   }
   ////////////////////
 
- long = '';
- lat = '';
+  private _long = ''
+  get long(): string {
+    return this._long
+  }
+  set long(value: string) {
+    this._long = value
+  }
+  private _lat = ''
+  get lat(): string {
+    return this._lat
+  }
+  set lat(value: string) {
+    this._lat = value
+  }
 
   // * Validations
   get validLat(): Boolean {
