+++ app/lucy/src/app/services/api.service.ts
@@ -294,19 +294,19 @@ export class ApiService {
   private async handleError(error: APIError): Promise<APIRequestResult> {
     switch (error.error.status) {
       case 401:
-        console.log(`Error 401 received, refreshing`);
+        // console.log(`Error 401 received, refreshing`);
         return await this.hendleErrorDescision(error, await this.decideOn401(error));
       case 404:
-        console.log(`Error 404 received: Resource is not Available`);
+        // console.log(`Error 404 received: Resource is not Available`);
         return await this.hendleErrorDescision(error, APIErrorDescision.Stop);
       case 500:
-        console.log(`Error 500 received: Resource is not Available`);
+        // console.log(`Error 500 received: Resource is not Available`);
         return await this.hendleErrorDescision(error, APIErrorDescision.Stop);
       case 422:
-          console.log(`Error 422 received: Unprocessable Entity`);
+          // console.log(`Error 422 received: Unprocessable Entity`);
           return await this.hendleErrorDescision(error, APIErrorDescision.Stop);
       default:
-        console.log(`ERRPR CASE NOT HANDLED.\n Error Code received: ${error.error.status}\nObject:`);
+        // console.log(`ERRPR CASE NOT HANDLED.\n Error Code received: ${error.error.status}\nObject:`);
         console.dir(error);
         return await this.hendleErrorDescision(error, APIErrorDescision.Stop);
     }
